added first draft to get ros graph
